Added a QueryMetadata class.
This new class stores the metadata (i.e. headers) of a query, and lets
users display the metadata as a table, look at the relevant metrics and
resource descriptors, and at the individual resource and metric labels.

The ResourceDescriptors class was updated to accept a list of resource
types for filtering.